+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/metrics/impl/visitors/CycloPathUnawareOperationVisitor.java
@@ -2,7 +2,7 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.lang.java.metrics.metrics.visitors;
+package net.sourceforge.pmd.lang.java.metrics.impl.visitors;
 
 import org.apache.commons.lang3.mutable.MutableInt;
 
@@ -23,7 +23,7 @@ import net.sourceforge.pmd.lang.java.ast.JavaParserVisitorAdapter;
  * Visitor calculating cyclo without counting boolean operators.
  *
  * @author Clément Fournier
- * @see net.sourceforge.pmd.lang.java.metrics.metrics.CycloMetric
+ * @see net.sourceforge.pmd.lang.java.metrics.impl.CycloMetric
  */
 public class CycloPathUnawareOperationVisitor extends JavaParserVisitorAdapter {
