+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/metrics/impl/visitors/JavaNcssVisitor.java
@@ -2,7 +2,7 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.lang.java.metrics.metrics.visitors;
+package net.sourceforge.pmd.lang.java.metrics.impl.visitors;
 
 import java.util.List;
 
@@ -17,7 +17,7 @@ import net.sourceforge.pmd.lang.java.ast.ASTPackageDeclaration;
  * JavaNcss compliant visitor for the calculation of Ncss.
  *
  * @author Clément Fournier
- * @see net.sourceforge.pmd.lang.java.metrics.metrics.NcssMetric
+ * @see net.sourceforge.pmd.lang.java.metrics.impl.NcssMetric
  */
 public class JavaNcssVisitor extends DefaultNcssVisitor {
