+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/metrics/impl/visitors/StandardCycloVisitor.java
@@ -2,7 +2,7 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.lang.java.metrics.metrics.visitors;
+package net.sourceforge.pmd.lang.java.metrics.impl.visitors;
 
 import org.apache.commons.lang3.mutable.MutableInt;
 
@@ -21,7 +21,7 @@ import net.sourceforge.pmd.lang.java.rule.codesize.NPathComplexityRule;
  * Calculates CYCLO following the standard definition.
  *
  * @author Clément Fournier
- * @see net.sourceforge.pmd.lang.java.metrics.metrics.CycloMetric
+ * @see net.sourceforge.pmd.lang.java.metrics.impl.CycloMetric
  */
 public class StandardCycloVisitor extends CycloPathUnawareOperationVisitor {
