+++ pmd-java/src/test/java/net/sourceforge/pmd/lang/java/metrics/DataStructureTest.java
@@ -29,8 +29,8 @@ import net.sourceforge.pmd.lang.java.ast.JavaParserVisitorReducedAdapter;
 import net.sourceforge.pmd.lang.java.ast.QualifiedName;
 import net.sourceforge.pmd.lang.java.metrics.api.JavaClassMetricKey;
 import net.sourceforge.pmd.lang.java.metrics.api.JavaOperationMetricKey;
-import net.sourceforge.pmd.lang.java.metrics.metrics.AbstractJavaClassMetric;
-import net.sourceforge.pmd.lang.java.metrics.metrics.AbstractJavaOperationMetric;
+import net.sourceforge.pmd.lang.java.metrics.impl.AbstractJavaClassMetric;
+import net.sourceforge.pmd.lang.java.metrics.impl.AbstractJavaOperationMetric;
 import net.sourceforge.pmd.lang.java.metrics.signature.FieldSigMask;
 import net.sourceforge.pmd.lang.java.metrics.signature.FieldSignature;
 import net.sourceforge.pmd.lang.java.metrics.signature.OperationSigMask;
