+++ pmd-java/src/test/resources/net/sourceforge/pmd/lang/java/metrics/impl/xml/WmcTest.xml
