+++ pmd-java/src/test/resources/rulesets/java/metrics_test.xml
@@ -11,25 +11,25 @@
 
     <rule name="CycloTest"
           message = "''{0}'' has value {1}."
-          class="net.sourceforge.pmd.lang.java.metrics.metrics.CycloTestRule"
+          class="net.sourceforge.pmd.lang.java.metrics.impl.CycloTestRule"
           metrics="true">
     </rule>
 
     <rule name="NcssTest"
           message = "''{0}'' has value {1}."
-          class="net.sourceforge.pmd.lang.java.metrics.metrics.NcssTestRule"
+          class="net.sourceforge.pmd.lang.java.metrics.impl.NcssTestRule"
           metrics="true">
     </rule>
 
     <rule name="WmcTest"
           message = "''{0}'' has value {1}."
-          class="net.sourceforge.pmd.lang.java.metrics.metrics.WmcTestRule"
+          class="net.sourceforge.pmd.lang.java.metrics.impl.WmcTestRule"
           metrics="true">
     </rule>
 
     <rule name="LocTest"
           message = "''{0}'' has value {1}."
-          class="net.sourceforge.pmd.lang.java.metrics.metrics.LocTestRule"
+          class="net.sourceforge.pmd.lang.java.metrics.impl.LocTestRule"
           metrics="true">
     </rule>
